requirements.txt
openai==0.28.1
python-dotenv==1.0.1
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
pyaudio==0.2.13
pydub==0.25.1
tqdm==4.66.2 